Trying to fix the type problem with chron_disease on arc

+++ minos/minosPipeline/RunPipeline.py
@@ -234,7 +234,7 @@ def type_check(data):
     data['S7_physical_health'] = data['S7_physical_health'].astype(int)
     data['nutrition_quality_diff'] = data['nutrition_quality_diff'].astype(int)
     data['neighbourhood_safety'] = data['neighbourhood_safety'].astype(int)
-    data['chron_disease'] = data['chron_disease'].astype(int)
+    data['chron_disease'] = data['chron_disease'].astype(float)
 
     return data
